% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_scatterplot.R
\name{vis_scatterplot}
\alias{vis_scatterplot}
\title{Plot a simple scatterplot}
\usage{
vis_scatterplot(data, x, y)
}
\arguments{
\item{data}{tibble Data to be plotted}

\item{x}{variable to be plotted on x-axis}

\item{y}{vairable to be plotted on y-axis}
}
\value{
A ggplot2 object
}
\description{
Plot a simple scatterplot
}
